pyproject.toml
[tool.poetry]
name = "web3-voting"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "web3_voting"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
eth-ape = "^0.6.15"
ape-alchemy = "^0.6.2"
ape-solidity= "^0.6.7"
ape-foundry = "^0.6.14"
flask = "2.3.2"
moralis = "0.1.37"
flask_cors = "^4.0.0"
python-dotenv = "1.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
